+++ insights/dashboards/usecases/conversations_dashboard_creation.py
@@ -12,28 +12,23 @@ CONVERSATIONS_DASHBOARD_WIDGETS = [
 
 class CreateConversationsDashboard:
     def create_dashboard(self, project):
-        if Dashboard.objects.filter(
-            project=project, name=CONVERSATIONS_DASHBOARD_NAME
-        ).exists():
-            raise Exception("Conversation dashboard already exists for this project")
-
-        with transaction.atomic():
-            dashboard = Dashboard.objects.create(
-                project=project,
-                name=CONVERSATIONS_DASHBOARD_NAME,
-                description="Conversations dashboard",
-                is_default=False,
-                grid=[0, 0],
-                is_deletable=False,
-                is_editable=True,
-                config={
+        dashboard, created = Dashboard.objects.get_or_create(
+            project=project,
+            name=CONVERSATIONS_DASHBOARD_NAME,
+            defaults={
+                "description": "Conversations dashboard",
+                "is_default": False,
+                "grid": [0, 0],
+                "is_deletable": False,
+                "is_editable": True,
+                "config": {
                     "type": "conversational",
                     "show_tool_result": True,
                     "show_agent_invocation": True,
                 },
-            )
-            self._create_widgets(dashboard)
-
+            },
+        )
+        self._create_widgets(dashboard)
         return dashboard
 
     def _create_widgets(self, dashboard):